:basketball:

+++ lintcode/HashFunction.java
@@ -38,7 +38,7 @@ public class HashFunction {
     public int hashCode(char[] key, int HASH_SIZE) {
         long hash = 0;
         for (int c : key) {
-            hash = (33 * hash % HASH_SIZE + (int) c % HASH_SIZE) % HASH_SIZE;
+            hash = (33 * hash % HASH_SIZE + c % HASH_SIZE) % HASH_SIZE;
         }
         return (int) hash;
     }
@@ -48,7 +48,7 @@ public class HashFunction {
     public int hashCode(char[] key, int HASH_SIZE) {
         long hash = 0;
         for (int i = 0; i < key.length; ++i) {
-            hash = (hash + (int) key[i] * powMod(33, key.length - 1 - i, 
+            hash = (hash + key[i] * powMod(33, key.length - 1 - i, 
                         HASH_SIZE) % HASH_SIZE) % HASH_SIZE;
         }
         return (int) hash;
